day7: this was straighforward brute-force.
Interestingly it seems to be very slightly faster to have my
initial enum solution than the [Fn(usize, usize) -> usize] version.

This is confusing to me because the 2nd one is branch-less, right ?
I don't think the branch predictor should be able to do a good job here.
Very surprising.